Moving version path helpers to their own lib/ files
