+++ src/lib/get-version-from-path.ts
@@ -1,4 +1,4 @@
-const REGEX = /^v\d+\.\d+\.(\d+|\w+)$/i
+import { VERSION_IN_PATH_REGEX } from 'constants/version-path'
 
 /**
  * Extract the version from a path string, or an array of path segments
@@ -6,7 +6,7 @@ const REGEX = /^v\d+\.\d+\.(\d+|\w+)$/i
 export function getVersionFromPath(path: string): string | undefined {
   const pathSegments = path.split('/')
 
-  const version = pathSegments.find((el) => REGEX.test(el))
+  const version = pathSegments.find((el) => VERSION_IN_PATH_REGEX.test(el))
 
   return version
 }
